new tests, refactor, and move languagebind cache_dir to standard location
